requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
astunparse=1.6.3=pypi_0
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.10.14=0
cached-property=1.5.2=pypi_0
cachetools=5.0.0=pypi_0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
charset-normalizer=2.0.12=pypi_0
click=7.1.2=py_0
cryptography=3.1.1=py37h7a1dbc1_0
flask=1.1.2=py_0
flatbuffers=2.0=pypi_0
freetype=2.10.4=hd328e21_0
gast=0.5.3=pypi_0
google-auth=2.6.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.44.0=pypi_0
h5py=3.6.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
itsdangerous=1.1.0=py37_0
jinja2=2.11.2=py_0
jpeg=9b=vc14h4d7706e_1
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
libclang=13.0.0=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=hf4a77e7_3
markdown=3.3.6=pypi_0
markupsafe=1.1.1=py37hfa6e2cd_1
numpy=1.21.5=pypi_0
oauthlib=3.2.0=pypi_0
olefile=0.46=py37_0
openssl=1.1.1n=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
pillow=8.0.0=py37hca74424_0
pip=21.2.4=py37haa95532_0
protobuf=3.20.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pyopenssl=19.1.0=py_1
pysocks=1.7.1=py37_1
python=3.7.13=h6244533_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
setuptools=58.0.4=py37haa95532_0
six=1.16.0=pypi_0
sqlite=3.38.2=h2bbff1b_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.0=pypi_0
tensorflow-io-gcs-filesystem=0.24.0=pypi_0
termcolor=1.1.0=pypi_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
tk=8.6.10=he774522_0
typing-extensions=4.1.1=pypi_0
urllib3=1.26.9=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37haa95532_2
wrapt=1.14.0=pypi_0
xz=5.2.5=h62dcd97_0
zipp=3.8.0=pypi_0
zlib=1.2.11=vc14h1cdd9ab_1
zstd=1.4.4=ha9fde0e_3
